Fix round 1: ticket_number vira BigInteger e cobre fronteira de paginacao
Alinha notifications.ticket_number com o tipo de origem (TicketModel.number,
BigInteger) na mesma migration 0009_notifications, sem revisao nova, ja que
nao ha dado gravado ainda. Adiciona teste de paginacao com per_page=2 sobre
3 notificacoes para exercitar offset/limit alem da primeira pagina, achado
do fix round 1 da revisao.

+++ backend/migrations/tenant/versions/0009_notifications.py
@@ -29,7 +29,9 @@ def upgrade() -> None:
         sa.Column("id", sa.Uuid(), primary_key=True),
         sa.Column("user_id", sa.Uuid(), nullable=False),
         sa.Column("ticket_id", sa.Uuid(), nullable=False),
-        sa.Column("ticket_number", sa.Integer(), nullable=False),
+        # BigInteger para espelhar tickets.number (0003_tickets), a coluna de
+        # origem que esta desnormalizada aqui.
+        sa.Column("ticket_number", sa.BigInteger(), nullable=False),
         sa.Column("type", sa.String(20), nullable=False),
         sa.Column("title", sa.String(200), nullable=False),
         sa.Column("snippet", sa.String(200), nullable=True),
